Upgrade vagrant environment to use wheezy
Mostly came out of the fact some Ruby gems didn't work on the Ruby
version distributed with squeeze, and that's now two versions behind
the current Debian, so upgrading to wheezy seemed worthwhile. :-)

I didn't upgrade to Jessie because the MariaDB packages don't have
any support for the current Debian version yet.

On a side note, MariaDB 10.0 seemed to work fine in a clustered state,
but I've intentionally left it at 5.5, since we only support the
previous stable version.